requirements.txt
langchain==0.1.15
transformers==4.39.3
sentence-transformers
accelerate
faiss-cpu
streamlit==1.33.0
langchain-fireworks
einops
langchain-together
htbuilder
python-dotenv
